requirements.txt
streamlit
transformers
timm 
einops
pillow
streamlit-extras
langchain
pyvips